+++ email/src/folder/mod.rs
@@ -1,4 +1,9 @@
-//! Module dedicated to folders (mailboxes).
+//! Module dedicated to email folders.
+//!
+//! This module contains [Folder] and [Folders] representations.
+//!
+//! You also have everything you need to [synchronize](sync) a remote
+//! folder with a local one.
 
 pub mod sync;
 
@@ -12,7 +17,33 @@ pub use self::sync::{
     FolderSyncPatchManager, FolderSyncPatches, FolderSyncStrategy,
 };
 
-/// Represents the list of folders.
+/// The email folder.
+///
+/// A folder is an email container. Depending on the [crate::Backend],
+/// the folder is an alias for a mailbox (IMAP) or a directory
+/// (Maildir/Notmuch).
+#[derive(Clone, Debug, Default, Eq, Hash)]
+pub struct Folder {
+    /// The folder name.
+    pub name: String,
+
+    /// The folder description.
+    pub desc: String,
+}
+
+impl PartialEq for Folder {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+    }
+}
+
+impl fmt::Display for Folder {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.name)
+    }
+}
+
+/// The list of folders.
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct Folders(Vec<Folder>);
 
@@ -35,28 +66,5 @@ impl FromIterator<Folder> for Folders {
         let mut folders = Folders::default();
         folders.extend(iter);
         folders
-    }
-}
-
-/// Represents the folder.
-#[derive(Clone, Debug, Default, Eq, Hash)]
-pub struct Folder {
-    /// Represents the folder hierarchie delimiter.
-    pub delim: String,
-    /// Represents the folder name.
-    pub name: String,
-    /// Represents the folder description.
-    pub desc: String,
-}
-
-impl PartialEq for Folder {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-    }
-}
-
-impl fmt::Display for Folder {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.name)
     }
 }